refactor(checkbox): move checkbox to its folder with theme interface
#443 #234
