+++ src/lib/components/FormControls/Label/Label.spec.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 import { HiGlobe, HiLockClosed } from 'react-icons/hi';
 import { describe, expect, it } from 'vitest';
 import { Button } from '../../Button';
-import { Checkbox } from '../Checkbox';
+import { Checkbox } from '../../Checkbox';
 import { FileInput } from '../FileInput';
 import { Radio } from '../Radio';
 import { Select } from '../Select';
