+++ src/lib/theme/default.ts
@@ -363,6 +363,9 @@ const theme: FlowbiteTheme = {
       span: 'self-center whitespace-nowrap text-2xl font-semibold text-gray-800 dark:text-white',
     },
   },
+  checkbox: {
+    base: 'h-4 w-4 rounded border border-gray-300 bg-gray-100 focus:ring-2 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:ring-offset-gray-800 dark:focus:ring-blue-600',
+  },
   formControls: {
     helperText: {
       base: 'mt-2 text-sm',
@@ -387,9 +390,6 @@ const theme: FlowbiteTheme = {
     },
     radio: {
       base: 'h-4 w-4 border border-gray-300 focus:ring-2 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:focus:bg-blue-600 dark:focus:ring-blue-600',
-    },
-    checkbox: {
-      base: 'h-4 w-4 rounded border border-gray-300 bg-gray-100 focus:ring-2 focus:ring-blue-500 dark:border-gray-600 dark:bg-gray-700 dark:ring-offset-gray-800 dark:focus:ring-blue-600',
     },
     textInput: {
       base: 'flex',
